Extract abstract base test classes and fix CS violations
ElasticSearchOfferQueryBuilderTest and ElasticSearchOrganizerQueryBuilderTest
were made non-final to allow the ES8 test classes to inherit their test suite.
Extract all test methods into abstract base classes so both the ES5 and ES8
concrete test classes can be final, satisfying the final_class CS rule.
Also fixes two double blank lines after namespace declarations.
